TTS: add translation.
Play flip sound only when automatic switch to next article.
Deactivate some logs.

+++ app/src/main/java/fr/gaulupeau/apps/Poche/tts/TextInterface.java
@@ -12,6 +12,6 @@ public interface TextInterface {
     boolean rewind();
     boolean skipToNext();
     boolean skipToPrevious();
-    float   getTime();
-    float   getTotalDuration();
+    long   getTime();
+    long   getTotalDuration();
 }
